moonfall: Fix ddcutil crashes
ddcutil was quiting violently with a coredump when ran from a udev rule.
I am not in the mood to debug why, but running it from a systemd oneshot
service seems to be a fine solution.

Additionally changed the trigger device for (de)windowsification to the
mouse, but that's purely a vibe-based assesment, just feels better.