workstations: Switch back to greetd
SDDM requires that sway is installed globally, which leads to problems
with swayidle missing env variables (WAYLAND_DISPLAY). This probably can
be fixed, but I don't really care which display manager is in use, as
long as it works :).
Previous mysterious issues with greetd seem to be related to not
switching the system profile to the new version in updater service.